+++ tests/Integration/DatabaseTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Kreait\Firebase\Tests\Integration;
 
+use Beste\Json;
 use Kreait\Firebase\Exception\Database\DatabaseNotFound;
 use Kreait\Firebase\Factory;
 
@@ -21,7 +22,7 @@ final class DatabaseTest extends DatabaseTestCase
             $this->markTestSkipped('The RTDB emulator creates databases if they don\'t exist');
         }
 
-        $credentials = self::$serviceAccount->asArray();
+        $credentials = Json::decode(self::$serviceAccount, true);
         $credentials['project_id'] = 'non-existing';
 
         $this->expectException(DatabaseNotFound::class);